requirements.txt
flask==2.0.1
pyserial==3.5
bleak==0.19.0
asyncio==3.4.3 